Integrate all CRUD functions for reg requests
Add edit functionality and clean up how the requests delete. Also made
it so the status of requests are updated from unsubmitted to submitted
on form submission. Requests are still being made when the new form is
loaded, so one is being created with every reload, which will be fixed
in a later version.