Add response, fix status and creator formatting
Made it so instead of only having a student description, there is
also a response section from the admin. This is allowed by the use of
the formatted_descr section of the reg_request, which is filled with the
admin response. Also shows creator name instead of id and a human
readable status over the standard kuali API designations.